Server now updates the currency memory on user request as well.
In exchange controller we check if the last update date is today. If yes we do nothing, otherwise we call the database update function with the date of the latest update.
The database update function sends a promise request to the xml database, parses it and if the latest update equals the latest data in the XML we do nothing otherwise we insert the new data to the database.